shared_ptr<int> as function evaluation counter
the objective function in class `Optimizer` is a runtime-generated
lambda function, prev version use a `static int` variable as a counter
to count the number of function evaluation. But that works only if
function `Optimizer::gen_opt_func` is called only once. Although it is true
currently, it's still unsafe for future development.

Closure might be a better solution, but it seems c++ doesn't support
Perl-like(or js-like) closure, so, I have to use a `shared_ptr`